+++ src/main/java/com/raoulvdberge/refinedstorage/proxy/ProxyCommon.java
@@ -106,7 +106,8 @@ public class ProxyCommon {
                 buf -> new CraftingMonitorElementItemRender(StackUtils.readItemStack(buf), buf.readInt(), buf.readInt(),
                         buf.readInt(), buf.readInt()));
         API.instance().getCraftingMonitorElementRegistry().add(CraftingMonitorElementFluidRender.ID,
-                buf -> new CraftingMonitorElementFluidRender(FluidStack.loadFluidStackFromNBT(ByteBufUtils.readTag(buf)),
+                buf -> new CraftingMonitorElementFluidRender(
+                        FluidStack.loadFluidStackFromNBT(ByteBufUtils.readTag(buf)),
                         buf.readInt(), buf.readInt(), buf.readInt(), buf.readInt()));
         API.instance().getCraftingMonitorElementRegistry().add(CraftingMonitorElementError.ID, buf -> {
             String id = ByteBufUtils.readUTF8String(buf);
@@ -175,8 +176,11 @@ public class ProxyCommon {
         RS.INSTANCE.network
                 .registerMessage(MessageGridItemInsertHeld.class, MessageGridItemInsertHeld.class, id++, Side.SERVER);
         RS.INSTANCE.network.registerMessage(MessageGridItemPull.class, MessageGridItemPull.class, id++, Side.SERVER);
-        RS.INSTANCE.network.registerMessage(MessageGridItemScroll.class, MessageGridItemScroll.class, id++, Side.SERVER);
-        RS.INSTANCE.network.registerMessage(MessageGridItemInventoryScroll.class, MessageGridItemInventoryScroll.class, id++, Side.SERVER);
+        RS.INSTANCE.network
+                .registerMessage(MessageGridItemScroll.class, MessageGridItemScroll.class, id++, Side.SERVER);
+        RS.INSTANCE.network
+                .registerMessage(MessageGridItemInventoryScroll.class, MessageGridItemInventoryScroll.class, id++,
+                        Side.SERVER);
         RS.INSTANCE.network.registerMessage(MessageGridClear.class, MessageGridClear.class, id++, Side.SERVER);
         RS.INSTANCE.network.registerMessage(MessageGridTransfer.class, MessageGridTransfer.class, id++, Side.SERVER);
         RS.INSTANCE.network
@@ -194,6 +198,9 @@ public class ProxyCommon {
         RS.INSTANCE.network
                 .registerMessage(MessageGridItemUpdate.class, MessageGridItemUpdate.class, id++, Side.CLIENT);
         RS.INSTANCE.network.registerMessage(MessageGridItemDelta.class, MessageGridItemDelta.class, id++, Side.CLIENT);
+        RS.INSTANCE.network
+                .registerMessage(MessageGridItemInsertInventory.class, MessageGridItemInsertInventory.class, id++,
+                        Side.SERVER);
         RS.INSTANCE.network
                 .registerMessage(MessageGridFluidUpdate.class, MessageGridFluidUpdate.class, id++, Side.CLIENT);
         RS.INSTANCE.network