+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 2.0.5
 - Fixed crafting grid duplicating items when not connected to a network
+- Allow using space to insert whole inventory
 
 ### 2.0.4
 - The crafting grid now automatically refills buckets